requirements.txt
# Web Framework
flask==2.2.3
werkzeug==2.2.3
gunicorn==21.2.0

# AI e Machine Learning
openai==1.3.0
assemblyai==0.17.0

# Processamento de Áudio
pydub==0.25.1

# Banco de Dados
supabase==2.0.0

# Google APIs
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
google-api-python-client==2.100.0

# Agno Framework
agno==0.1.0

# Utilitários
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.5

# Logging e Debugging
coloredlogs==15.0.1
